+++ tests/performance/reduce_vs_array_reduce.php
@@ -5,13 +5,16 @@ use Symfony\Component\Console\Helper\Table;
 
 include_once __DIR__ . "/../../vendor/autoload.php";
 
+const NUMBER_OF_ITEMS = 10000;
+const REPEAT_COUNT = 10;
+
 function getIntegerReport()
 {
     $arrayMapDeltas = 0.0;
     $collectionMapDeltas = 0.0;
     $fixtureProvider = function () {
         $array = [];
-        for ($i = 0; $i < 1000; $i++) {
+        for ($i = 0; $i < NUMBER_OF_ITEMS; $i++) {
             $array[] = $i;
         }
 
@@ -21,7 +24,7 @@ function getIntegerReport()
         return $item + $temp;
     };
 
-    for ($j = 0; $j < 10; $j++) {
+    for ($j = 0; $j < REPEAT_COUNT; $j++) {
         $array = $fixtureProvider();
 
 
@@ -36,9 +39,9 @@ function getIntegerReport()
     }
 
     return [
-        'name' => 'array_reduce vs Collection::reduce on 1000 integers (addition)',
-        'native' => $arrayMapDeltas / 10.0,
-        'collection' => $collectionMapDeltas / 10.0
+        'name' => 'array_reduce vs Collection::reduce on ' . NUMBER_OF_ITEMS . ' integers (addition)',
+        'native' => (float) $arrayMapDeltas / REPEAT_COUNT,
+        'collection' => (float) $collectionMapDeltas / REPEAT_COUNT
     ];
 }
 
@@ -48,7 +51,7 @@ function getStringReport()
     $collectionMapDeltas = 0.0;
     $fixtureProvider = function () {
         $array = [];
-        for ($i = 0; $i < 1000; $i++) {
+        for ($i = 0; $i < NUMBER_OF_ITEMS; $i++) {
             $array[] = $i . 'asd';
         }
 
@@ -58,7 +61,7 @@ function getStringReport()
         return $temp . $item;
     };
 
-    for ($j = 0; $j < 10; $j++) {
+    for ($j = 0; $j < REPEAT_COUNT; $j++) {
         $array = $fixtureProvider();
         $arrayMapStart = microtime(true);
         array_reduce($array, $mapper, '');
@@ -72,9 +75,9 @@ function getStringReport()
     }
 
     return [
-        'name' => 'array_reduce vs Collection::reduce on 1000 strings (concatenation)',
-        'native' => $arrayMapDeltas / 10.0,
-        'collection' => $collectionMapDeltas / 10.0
+        'name' => 'array_reduce vs Collection::reduce on ' . NUMBER_OF_ITEMS . ' strings (concatenation)',
+        'native' => (float) $arrayMapDeltas / REPEAT_COUNT,
+        'collection' => (float) $collectionMapDeltas / REPEAT_COUNT
     ];
 }
 
@@ -84,7 +87,7 @@ function getObjectReport()
     $collectionMapDeltas = 0.0;
     $fixtureProvider = function () {
         $array = [];
-        for ($i = 0; $i < 1000; $i++) {
+        for ($i = 0; $i < NUMBER_OF_ITEMS; $i++) {
             $c = new stdClass();
             $c->asd = 1;
             $array[] = $c;
@@ -96,7 +99,7 @@ function getObjectReport()
         return $temp + $item->asd;
     };
 
-    for ($j = 0; $j < 10; $j++) {
+    for ($j = 0; $j < REPEAT_COUNT; $j++) {
         $array = $fixtureProvider();
         $arrayMapStart = microtime(true);
         array_reduce($array, $mapper, 0);
@@ -110,9 +113,9 @@ function getObjectReport()
     }
 
     return [
-        'name' => 'array_reduce vs Collection::reduce on 1000 object (object to field value)',
-        'native' => $arrayMapDeltas / 10.0,
-        'collection' => $collectionMapDeltas / 10.0
+        'name' => 'array_reduce vs Collection::reduce on ' . NUMBER_OF_ITEMS . ' object (object to field value)',
+        'native' => (float) $arrayMapDeltas / REPEAT_COUNT,
+        'collection' => (float) $collectionMapDeltas / REPEAT_COUNT
     ];
 }
 
@@ -122,7 +125,7 @@ function getComplexOperationReport()
     $collectionMapDeltas = 0.0;
     $fixtureProvider = function () {
         $array = [];
-        for ($i = 0; $i < 1000; $i++) {
+        for ($i = 0; $i < NUMBER_OF_ITEMS; $i++) {
             $array[] = $i;
         }
 
@@ -137,7 +140,7 @@ function getComplexOperationReport()
         return $temp + $result;
     };
 
-    for ($j = 0; $j < 10; $j++) {
+    for ($j = 0; $j < REPEAT_COUNT; $j++) {
         $array = $fixtureProvider();
         $arrayMapStart = microtime(true);
         array_reduce($array, $mapper, 0);
@@ -151,9 +154,9 @@ function getComplexOperationReport()
     }
 
     return [
-        'name' => 'array_reduce vs Collection::reduce for 1000 integers n, counting sum(0, n) the naive way',
-        'native' => $arrayMapDeltas / 10.0,
-        'collection' => $collectionMapDeltas / 10.0
+        'name' => 'array_reduce vs Collection::reduce for ' . NUMBER_OF_ITEMS . ' integers n, counting sum(0, n) the naive way',
+        'native' => (float) $arrayMapDeltas / REPEAT_COUNT,
+        'collection' => (float) $collectionMapDeltas / REPEAT_COUNT
     ];
 }
 
@@ -163,7 +166,7 @@ function getHashReport()
     $collectionMapDeltas = 0.0;
     $fixtureProvider = function () {
         $array = [];
-        for ($i = 0; $i < 1000; $i++) {
+        for ($i = 0; $i < NUMBER_OF_ITEMS; $i++) {
             $array[] = $i . 'asdf';
         }
 
@@ -173,7 +176,7 @@ function getHashReport()
         return md5($temp . $item);
     };
 
-    for ($j = 0; $j < 10; $j++) {
+    for ($j = 0; $j < REPEAT_COUNT; $j++) {
         $array = $fixtureProvider();
         $arrayMapStart = microtime(true);
         array_reduce($array, $mapper, '');
@@ -187,9 +190,9 @@ function getHashReport()
     }
 
     return [
-        'name' => 'array_reduce vs Collection::reduce on 1000 md5 invocations',
-        'native' => $arrayMapDeltas / 10.0,
-        'collection' => $collectionMapDeltas / 10.0
+        'name' => 'array_reduce vs Collection::reduce on ' . NUMBER_OF_ITEMS . ' md5 invocations',
+        'native' => (float) $arrayMapDeltas / REPEAT_COUNT,
+        'collection' => (float) $collectionMapDeltas / REPEAT_COUNT
     ];
 }