+++ tests/scenarios/FibonaccisSequenceTest.php
@@ -7,6 +7,9 @@ use PHPUnit_Framework_TestCase;
 
 class FibonaccisSequenceTest extends PHPUnit_Framework_TestCase
 {
+    /**
+     * Example generating first 5 values in fibonacci's sequence.
+     */
     public function testIt()
     {
         $result = Collection::iterate([1, 1], function ($v) {